backend/requirements.txt
# ============================================================
# ReplicarIA - Backend Dependencies
# ============================================================

# === Core Framework ===
fastapi==0.110.1
uvicorn[standard]>=0.25.0
gunicorn>=21.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# === AI/LLM Providers ===
anthropic>=0.34.0
openai>=1.0.0
tiktoken>=0.5.0

# === HTTP Clients ===
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0

# === Authentication & Security ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# === MongoDB (CRÍTICO - faltaba) ===
motor>=3.3.0
pymongo>=4.6.0

# === PostgreSQL ===
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
greenlet>=3.0.0

# === PDF Processing ===
PyPDF2>=3.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.17.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# === Document Processing ===
python-docx>=1.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# === File & Image Processing ===
aiofiles>=23.0.0
Pillow>=10.0.0
qrcode>=7.4.0

# === Testing (CRÍTICO - faltaba pytest-asyncio) ===
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# === Optional Services ===
redis>=5.0.0

# === Email (si se usa) ===
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
